HBX-1995: Rename package 'org.hibernate.tool.internal.dialect' to 'org.hibernate.tool.internal.reveng.dialect'
