+++ orm/src/main/java/org/hibernate/tool/internal/reveng/dialect/HSQLMetaDataDialect.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.internal.dialect;
+package org.hibernate.tool.internal.reveng.dialect;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
