+++ orm/src/main/java/org/hibernate/tool/internal/reveng/dialect/OracleMetaDataDialect.java
@@ -1,5 +1,5 @@
 
-package org.hibernate.tool.internal.dialect;
+package org.hibernate.tool.internal.reveng.dialect;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
