+++ test/common/src/main/java/org/hibernate/tool/cfg/DriverMetaData/TestCase.java
@@ -14,8 +14,8 @@ import org.hibernate.cfg.Environment;
 import org.hibernate.engine.jdbc.connections.spi.ConnectionProvider;
 import org.hibernate.service.ServiceRegistry;
 import org.hibernate.tool.api.reveng.RevengDialectFactory;
+import org.hibernate.tool.internal.reveng.dialect.JDBCMetaDataDialect;
 import org.hibernate.tool.api.reveng.RevengDialect;
-import org.hibernate.tool.internal.dialect.JDBCMetaDataDialect;
 import org.hibernate.tools.test.util.JUnitUtil;
 import org.hibernate.tools.test.util.JdbcUtil;
 import org.junit.After;
