+++ test/nodb/src/test/java/org/hibernate/tool/cfg/MetaDataDialectFactoryTest/TestCase.java
@@ -11,11 +11,11 @@ import org.hibernate.dialect.Oracle10gDialect;
 import org.hibernate.dialect.Oracle8iDialect;
 import org.hibernate.dialect.Oracle9iDialect;
 import org.hibernate.tool.api.reveng.RevengDialectFactory;
-import org.hibernate.tool.internal.dialect.H2MetaDataDialect;
-import org.hibernate.tool.internal.dialect.HSQLMetaDataDialect;
-import org.hibernate.tool.internal.dialect.JDBCMetaDataDialect;
-import org.hibernate.tool.internal.dialect.MySQLMetaDataDialect;
-import org.hibernate.tool.internal.dialect.OracleMetaDataDialect;
+import org.hibernate.tool.internal.reveng.dialect.H2MetaDataDialect;
+import org.hibernate.tool.internal.reveng.dialect.HSQLMetaDataDialect;
+import org.hibernate.tool.internal.reveng.dialect.JDBCMetaDataDialect;
+import org.hibernate.tool.internal.reveng.dialect.MySQLMetaDataDialect;
+import org.hibernate.tool.internal.reveng.dialect.OracleMetaDataDialect;
 import org.junit.Assert;
 import org.junit.Test;
