+++ test/oracle/src/main/java/org/hibernate/cfg/reveng/dialect/TestCase.java
@@ -13,7 +13,7 @@ import org.hibernate.cfg.Environment;
 import org.hibernate.engine.jdbc.connections.spi.ConnectionProvider;
 import org.hibernate.service.ServiceRegistry;
 import org.hibernate.tool.api.reveng.RevengDialect;
-import org.hibernate.tool.internal.dialect.OracleMetaDataDialect;
+import org.hibernate.tool.internal.reveng.dialect.OracleMetaDataDialect;
 import org.hibernate.tools.test.util.JdbcUtil;
 import org.junit.After;
 import org.junit.Before;
